rename datawrapper -> resource
