+++ pygfx/geometries/_base.py
@@ -1,10 +1,10 @@
 import numpy as np
 
-from ..objects._base import DataWrapperContainer
-from ..datawrappers import Buffer
+from ..objects._base import ResourceContainer
+from ..resources import Buffer
 
 
-class Geometry(DataWrapperContainer):
+class Geometry(ResourceContainer):
     """ A geometry represents the (input) data of mesh, line, or point
     geometry. It can include vertex positions, normals, colors, uvs,
     and custom data buffers. Face indices can be given using `index`.
