+++ pygfx/materials/_points.py
@@ -2,7 +2,7 @@ from pyshader import Struct, f32, vec4
 
 from ._base import Material
 from ..utils import array_from_shadertype
-from ..datawrappers import Buffer
+from ..resources import Buffer
 
 
 class PointsMaterial(Material):
