+++ pygfx/resources/__init__.py
