+++ pygfx/resources/_buffer.py
