+++ pygfx/resources/_texture.py
